Move rest of elements utils out of API
We should split this up better as metadata.kt has become a dumping ground
